HANZEanalytics/pyproject.toml
[project]
name = "hanzeanalytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cartopy>=0.25.0",
    "geopandas>=1.1.3",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "wget>=3.2",
]
[tool.uv]
package = true
